Move coordinator to separate file Fully type hinting
