+++ homeassistant/components/speedtestdotnet/__init__.py
@@ -2,34 +2,34 @@
 from __future__ import annotations
 
 from datetime import timedelta
-import logging
 
 import speedtest
 
 from homeassistant.config_entries import ConfigEntry
-from homeassistant.const import CONF_SCAN_INTERVAL, EVENT_HOMEASSISTANT_STARTED
-from homeassistant.core import CoreState, HomeAssistant
-from homeassistant.exceptions import ConfigEntryNotReady
-from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
-
-from .const import (
-    CONF_MANUAL,
-    CONF_SERVER_ID,
-    DEFAULT_SCAN_INTERVAL,
-    DEFAULT_SERVER,
-    DOMAIN,
-    PLATFORMS,
+from homeassistant.const import (
+    CONF_SCAN_INTERVAL,
+    EVENT_HOMEASSISTANT_STARTED,
+    Platform,
 )
+from homeassistant.core import CoreState, Event, HomeAssistant
+from homeassistant.exceptions import ConfigEntryNotReady
 
-_LOGGER = logging.getLogger(__name__)
+from .const import CONF_MANUAL, DEFAULT_SCAN_INTERVAL, DOMAIN
+from .coordinator import SpeedTestDataCoordinator
+
+PLATFORMS = [Platform.SENSOR]
 
 
 async def async_setup_entry(hass: HomeAssistant, config_entry: ConfigEntry) -> bool:
     """Set up the Speedtest.net component."""
-    coordinator = SpeedTestDataCoordinator(hass, config_entry)
-    await coordinator.async_setup()
+    api = speedtest.Speedtest(secure=True)
+    coordinator = SpeedTestDataCoordinator(hass, config_entry, api)
+    try:
+        await hass.async_add_executor_job(coordinator.update_servers)
+    except speedtest.SpeedtestException as err:
+        raise ConfigEntryNotReady from err
 
-    async def _enable_scheduled_speedtests(*_):
+    async def _enable_scheduled_speedtests(event: Event | None = None) -> None:
         """Activate the data update coordinator."""
         coordinator.update_interval = timedelta(
             minutes=config_entry.options.get(CONF_SCAN_INTERVAL, DEFAULT_SCAN_INTERVAL)
@@ -51,104 +51,28 @@ async def async_setup_entry(hass: HomeAssistant, config_entry: ConfigEntry) -> b
 
     await hass.config_entries.async_forward_entry_setups(config_entry, PLATFORMS)
 
+    config_entry.async_on_unload(
+        config_entry.add_update_listener(options_updated_listener)
+    )
+
     return True
 
 
 async def async_unload_entry(hass: HomeAssistant, config_entry: ConfigEntry) -> bool:
     """Unload SpeedTest Entry from config_entry."""
-    unload_ok = await hass.config_entries.async_unload_platforms(
+    if unload_ok := await hass.config_entries.async_unload_platforms(
         config_entry, PLATFORMS
-    )
-    if unload_ok:
+    ):
         hass.data.pop(DOMAIN)
     return unload_ok
 
 
-class SpeedTestDataCoordinator(DataUpdateCoordinator):
-    """Get the latest data from speedtest.net."""
-
-    def __init__(self, hass: HomeAssistant, config_entry: ConfigEntry) -> None:
-        """Initialize the data object."""
-        self.hass = hass
-        self.config_entry: ConfigEntry = config_entry
-        self.api: speedtest.Speedtest | None = None
-        self.servers: dict[str, dict] = {DEFAULT_SERVER: {}}
-        super().__init__(
-            self.hass,
-            _LOGGER,
-            name=DOMAIN,
-            update_method=self.async_update,
-        )
-
-    def initialize(self) -> None:
-        """Initialize speedtest api."""
-        self.api = speedtest.Speedtest(secure=True)
-        self.update_servers()
-
-    def update_servers(self):
-        """Update list of test servers."""
-        test_servers = self.api.get_servers()
-        test_servers_list = []
-        for servers in test_servers.values():
-            for server in servers:
-                test_servers_list.append(server)
-        for server in sorted(
-            test_servers_list,
-            key=lambda server: (
-                server["country"],
-                server["name"],
-                server["sponsor"],
-            ),
-        ):
-            self.servers[
-                f"{server['country']} - {server['sponsor']} - {server['name']}"
-            ] = server
-
-    def update_data(self):
-        """Get the latest data from speedtest.net."""
-        self.update_servers()
-        self.api.closest.clear()
-        if self.config_entry.options.get(CONF_SERVER_ID):
-            server_id = self.config_entry.options.get(CONF_SERVER_ID)
-            self.api.get_servers(servers=[server_id])
-
-        best_server = self.api.get_best_server()
-        _LOGGER.debug(
-            "Executing speedtest.net speed test with server_id: %s",
-            best_server["id"],
-        )
-        self.api.download()
-        self.api.upload()
-        return self.api.results.dict()
-
-    async def async_update(self) -> dict[str, str]:
-        """Update Speedtest data."""
-        try:
-            return await self.hass.async_add_executor_job(self.update_data)
-        except speedtest.NoMatchedServers as err:
-            raise UpdateFailed("Selected server is not found.") from err
-        except speedtest.SpeedtestException as err:
-            raise UpdateFailed(err) from err
-
-    async def async_setup(self) -> None:
-        """Set up SpeedTest."""
-        try:
-            await self.hass.async_add_executor_job(self.initialize)
-        except speedtest.SpeedtestException as err:
-            raise ConfigEntryNotReady from err
-
-        self.config_entry.async_on_unload(
-            self.config_entry.add_update_listener(options_updated_listener)
-        )
-
-
 async def options_updated_listener(hass: HomeAssistant, entry: ConfigEntry) -> None:
     """Handle options update."""
+    coordinator: SpeedTestDataCoordinator = hass.data[DOMAIN]
     if entry.options[CONF_MANUAL]:
-        hass.data[DOMAIN].update_interval = None
+        coordinator.update_interval = None
         return
 
-    hass.data[DOMAIN].update_interval = timedelta(
-        minutes=entry.options[CONF_SCAN_INTERVAL]
-    )
-    await hass.data[DOMAIN].async_request_refresh()
+    coordinator.update_interval = timedelta(minutes=entry.options[CONF_SCAN_INTERVAL])
+    await coordinator.async_request_refresh()
