+++ homeassistant/components/speedtestdotnet/sensor.py
@@ -1,10 +1,17 @@
 """Support for Speedtest.net internet speed testing sensor."""
 from __future__ import annotations
 
+from collections.abc import Callable
+from dataclasses import dataclass
 from typing import Any, cast
 
-from homeassistant.components.sensor import SensorEntity
+from homeassistant.components.sensor import (
+    SensorEntity,
+    SensorEntityDescription,
+    SensorStateClass,
+)
 from homeassistant.config_entries import ConfigEntry
+from homeassistant.const import TIME_MILLISECONDS, UnitOfDataRate
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.device_registry import DeviceEntryType
 from homeassistant.helpers.entity import DeviceInfo
@@ -13,7 +20,6 @@ from homeassistant.helpers.restore_state import RestoreEntity
 from homeassistant.helpers.typing import StateType
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from . import SpeedTestDataCoordinator
 from .const import (
     ATTR_BYTES_RECEIVED,
     ATTR_BYTES_SENT,
@@ -24,8 +30,38 @@ from .const import (
     DEFAULT_NAME,
     DOMAIN,
     ICON,
-    SENSOR_TYPES,
-    SpeedtestSensorEntityDescription,
+)
+from .coordinator import SpeedTestDataCoordinator
+
+
+@dataclass
+class SpeedtestSensorEntityDescription(SensorEntityDescription):
+    """Class describing Speedtest sensor entities."""
+
+    value: Callable = round
+
+
+SENSOR_TYPES: tuple[SpeedtestSensorEntityDescription, ...] = (
+    SpeedtestSensorEntityDescription(
+        key="ping",
+        name="Ping",
+        native_unit_of_measurement=TIME_MILLISECONDS,
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    SpeedtestSensorEntityDescription(
+        key="download",
+        name="Download",
+        native_unit_of_measurement=UnitOfDataRate.MEGABITS_PER_SECOND,
+        state_class=SensorStateClass.MEASUREMENT,
+        value=lambda value: round(value / 10**6, 2),
+    ),
+    SpeedtestSensorEntityDescription(
+        key="upload",
+        name="Upload",
+        native_unit_of_measurement=UnitOfDataRate.MEGABITS_PER_SECOND,
+        state_class=SensorStateClass.MEASUREMENT,
+        value=lambda value: round(value / 10**6, 2),
+    ),
 )
